refactor search screen
