+++ src/data/queries/PassageQuery.js
@@ -1,7 +1,6 @@
 import { repository } from '../models/Models';
 
 export function getPassage(activeBook, activeChapter) {
-    console.log('aaa', repository)
     var filteredPassages = repository().objects('Passage').filtered(
         `book = "${activeBook.value}" AND chapter = "${activeChapter}"`
     );
@@ -10,6 +9,23 @@ export function getPassage(activeBook, activeChapter) {
         const verses = versesRaw.map(key => filteredPassages[key]);
         return verses;
     }
-    console.log('adsf', versesRaw)
     return filteredPassages;
+}
+
+export function searchPassage(searchText) {
+    let passages = repository().objects("Passage");
+    if (searchText.indexOf(" ") != -1) {
+        let splitWords = searchText.replace("  ", " ").split(" ");
+        splitWords = splitWords
+            .map(word => (word != "" ? `content CONTAINS[c] "${word}"` : ""))
+            .filter(word => word);
+        // query = `${query} ${splitWords}`;
+        var query = `${splitWords.join(" OR ")} AND type != "t"`;
+    } else {
+        query = `content CONTAINS[c] "${searchText}" AND type != "t"`;
+    }
+    let filteredPassages = passages.filtered(query).slice(0, 20);
+    const resultsRaw = Object.keys(filteredPassages);
+    const results = resultsRaw.map(key => filteredPassages[key]);
+    return results;
 }
