+++ src/screens/Search/SearchScreen.js
@@ -10,75 +10,52 @@ import {
   TextInput,
   StatusBar
 } from "react-native";
-const Realm = require("realm");
-import DrawerLayout from "react-native-drawer-layout";
-import LinearGradient from "react-native-linear-gradient";
+import PropTypes from 'prop-types';
 import Books from "../../constants/Books";
 import Icon from "react-native-vector-icons/Ionicons";
 import Highlighter from "react-native-highlight-words";
-
-const PassageSchema = {
-  name: "Passage",
-  primaryKey: "id",
-  properties: {
-    id: "string",
-    content: "string",
-    book: "string",
-    chapter: "int",
-    verse: "int",
-    type: "string"
-  }
-};
+import { COLOR } from "../../constants/constants";
+import { queries } from '../../data/queries/Queries';
+const { searchPassage } = queries
 
 class SearchScreen extends Component {
   _textInputSearch;
   state = {
     results: []
   };
+
   componentDidMount() {
     setTimeout(() => {
       this._textInputSearch.focus();
       if (this.props.searchText != "") this.searchPassage();
     }, 100);
   }
+
   searchPassage() {
     const { searchText } = this.props;
     if (searchText == "") return;
-    Realm.open({ schema: [PassageSchema], readOnly: true }).then(realm => {
-      let passages = realm.objects("Passage");
-      if (searchText.indexOf(" ") != -1) {
-        let splitWords = searchText.replace("  ", " ").split(" ");
-        splitWords = splitWords
-          .map(word => (word != "" ? `content CONTAINS[c] "${word}"` : ""))
-          .filter(word => word);
-        // query = `${query} ${splitWords}`;
-        var query = `${splitWords.join(" OR ")} AND type != "t"`;
-      } else {
-        var query = `content CONTAINS[c] "${searchText}" AND type != "t"`;
-      }
-      let filteredPassages = passages.filtered(query).slice(0, 20);
-      const resultsRaw = Object.keys(filteredPassages);
-      const results = resultsRaw.map(key => filteredPassages[key]);
-      console.log(resultsRaw);
-      this.setState({
-        results: results
-      });
-    });
+    const results = searchPassage(searchText)
+    this.setState({results})
   }
+
   _onHideSearch() {
     this.props.onHideSearch && this.props.onHideSearch();
   }
+
   _onChangeSearchText(searchText) {
     this.props.onChangeSearchText && this.props.onChangeSearchText(searchText);
     this.searchPassage();
   }
+
   _onJumpPassage(verse) {
     this.props.onJumpPassage && this.props.onJumpPassage(verse);
     this._onHideSearch();
   }
+
   _onClearSearchText() {
     this.props.onChangeSearchText && this.props.onChangeSearchText("");
   }
+
   render() {
     const { results } = this.state;
     const { searchText } = this.props;
@@ -130,7 +107,7 @@ class SearchScreen extends Component {
                     {verse.chapter}:{verse.verse}{" "}
                   </Text>
                   <Highlighter
-                    highlightStyle={{ backgroundColor: "#26405A" }}
+                    highlightStyle={styles.highLighter}
                     searchWords={searchText.split(" ")}
                     textToHighlight={verse.content}
                   />
@@ -146,23 +123,18 @@ class SearchScreen extends Component {
 
 const styles = StyleSheet.create({
   container: {
-    flex: 1
+    flex: 1,
+    backgroundColor: COLOR.primaryExtraDark
+  },
+  highLighter: {
+    backgroundColor: COLOR.primaryDark
   },
   toolbar: {
     paddingTop: Platform.OS == "ios" ? 20 : 0,
     flexDirection: "row",
     height: Platform.OS == "ios" ? 80 : 60,
     alignItems: "center",
-    backgroundColor: "#fff"
-  },
-  titleButton: {
-    flex: 1,
-    paddingHorizontal: 25
-  },
-  toolbarTitle: {
-    color: "#fff",
-    fontSize: 20,
-    fontWeight: "300"
+    backgroundColor: COLOR.white
   },
   innerScroll: {
     paddingVertical: 30
@@ -171,73 +143,10 @@ const styles = StyleSheet.create({
   actionButton: {
     paddingHorizontal: 20
   },
-  container: {
-    flex: 1,
-    backgroundColor: "#0D233A"
-  },
-  drawerWrapper: {
-    backgroundColor: "#1f364d",
-    flex: 1,
-    paddingTop: 25
-  },
   text: {
-    color: "#fff",
+    color: COLOR.white,
     lineHeight: 30,
     paddingHorizontal: 25
-  },
-  title: {
-    fontWeight: "900"
-  },
-  separator: {
-    fontWeight: "900",
-    color: "rgba(255,255,255,0.4)",
-    fontSize: 10,
-    marginVertical: 10,
-    marginLeft: 20
-  },
-  book: {
-    paddingHorizontal: 25,
-    paddingVertical: 20
-  },
-  bookText: {
-    color: "#fff",
-    fontWeight: "300"
-  },
-  bookActive: {
-    backgroundColor: "#26405A"
-  },
-  chapterSelector: {
-    height: 60,
-    backgroundColor: "#26405A"
-  },
-  chapter: {
-    height: 60,
-    width: 60,
-    justifyContent: "center",
-    alignItems: "center"
-  },
-  chapterText: {
-    color: "#fff"
-  },
-  chapterActive: {
-    width: 40,
-    height: 40,
-    borderRadius: 20,
-    backgroundColor: "#fff",
-    justifyContent: "center",
-    alignItems: "center"
-  },
-  chapterTextActive: {
-    color: "#0D233A"
-  },
-  chapterScroll: {
-    paddingHorizontal: 20
-  },
-  verseNumber: {
-    fontWeight: "900",
-    paddingRight: 20,
-    color: "rgba(255,255,255,0.4)",
-    fontSize: 10
   },
   input: {
     flex: 1,
@@ -249,3 +158,10 @@ const styles = StyleSheet.create({
 });
 
 export default SearchScreen;
+
+SearchScreen.propTypes = {
+  searchText: PropTypes.string.isRequired,
+  onHideSearch: PropTypes.func.isRequired,
+  onChangeSearchText: PropTypes.func.isRequired,
+  onJumpPassage: PropTypes.func.isRequired
+}
